+++ src/ui/server/controllers/quote/buyer-country/index.ts
@@ -3,7 +3,7 @@ import { FIELD_IDS, ROUTES, TEMPLATES } from '../../../constants';
 import api from '../../../api';
 import { mapCountries } from '../../../helpers/mappings/map-countries';
 import singleInputPageVariables from '../../../helpers/single-input-page-variables';
-import { validation as generateValidationErrors } from './validation';
+import { validation as generateValidationErrors } from '../../../shared-validation/buyer-country';
 import isChangeRoute from '../../../helpers/is-change-route';
 import getCountryByName from '../../../helpers/get-country-by-name';
 import { canGetAQuoteOnline, canGetAQuoteByEmail, cannotGetAQuote } from '../../../helpers/country-support';
